chore: refactor python tests. no functional changes just moving into seperate files
