+++ integ-tests/python/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Integration tests for the BAML Python client."""
