+++ integ-tests/python/tests/providers/__init__.py
@@ -1,0 +1,1 @@
+"""Provider-specific integration tests for the BAML Python client."""
